cleanup clang-y static analysis
Some performance improvements.  eg. pass by ref.

Some paranoia pre-initialization of variables
which will be initialized later by indirection.

And some misc. cleanup.

+++ ioc/singlesourcehooks.cpp
@@ -75,7 +75,7 @@ void qReport(unsigned level) noexcept {
     try{
         if (auto pPvxsServer = pvxsServer.load()) {
             std::ostringstream strm;
-            Detailed D(strm, level);
+            Detailed D(strm, (int)level);
             strm << *pPvxsServer;
             printf("%s", strm.str().c_str());
         }
